pyproject.toml
[project]
name = "paido-core"
version = "0.1.0"
description = ""
authors = [
    {name = "Hugo Rodriguez",email = "83481444+huhero@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)"
]

[tool.poetry]
packages = [{include = "paido_core", from = "src"}]


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.5"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
taskipy = "^1.14.1"
factory-boy = "^3.3.3"
freezegun = "^1.5.1"
testcontainers = "^4.10.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev src/paido_core/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src/paido_core -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
